rag-implementation-for-banking-knowledge-base/requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
pinecone
openai
python-dotenv
langchain_pinecone
langchain_openai
pypdf